Responded to Easwar's comments
