+++ xds/internal/balancer/outlierdetection/callcounter.go
@@ -37,7 +37,9 @@ func newCallCounter() *callCounter {
 // callCounter has two buckets, which each count successful and failing RPC's.
 // The activeBucket is used to actively count any finished RPC's, and the
 // inactiveBucket is populated with this activeBucket's data every interval for
-// use by the Outlier Detection algorithm. Caller must hold b.mu.
+// use by the Outlier Detection algorithm.
+//
+// Caller must hold b.mu.
 type callCounter struct {
 	// activeBucket updates every time a call finishes (from picker passed to
 	// Client Conn), so protect pointer read with atomic load of unsafe.Pointer
@@ -54,7 +56,9 @@ func (cc *callCounter) clear() {
 // "When the timer triggers, the inactive bucket is zeroed and swapped with the
 // active bucket. Then the inactive bucket contains the number of successes and
 // failures since the last time the timer triggered. Those numbers are used to
-// evaluate the ejection criteria." - A50. Caller must hold b.mu.
+// evaluate the ejection criteria." - A50.
+//
+// Caller must hold b.mu.
 func (cc *callCounter) swap() {
 	ib := cc.inactiveBucket
 	*ib = bucket{}
